Update README.md
Tomato is one of the most widely consumed vegetables in Iran and the world but it is suffering from annual losses. The diseases such as Alternaria and pests such as Leafminer are threats to this agricultural product. These pests and diseases first damage the leaves of tomatoes, then the flowers, and finally its fruit. Therefore, the damage to the tomato tree can be greatly reduced by controlling the disease and the pest on the leaves in their early stages. It is difficult for farmers to distinguish Alternaria disease from Leafminear pest at the early and middle stages of their outbreak on tomato leave. Therefore, in this study, in these stages, 272 leaf images including 172 images of tomato leaves infected with both Alternaria and Leafminear and 100 healthy tomato leaves were prepared from the farm of Vali-e-Asr University of Rafsanjan. The image processing technique, neural networks, and Adaptive Network based Fuzzy Inference System (ANFIS) classifiers were used to diagnose Alternaria disease and Leafminear pest on tomato leaves. The results showed that ANFIS classifier achieved an accuracy of 84.71% in 2.5 seconds when working on CPU. It can be used as an effective method for early diagnosis of Alternaria and Leafminear on tomato leaves to prevent their spread in the farm and greenhouse. This study also suggests a high classification accuracy for an intelligent selection of pixel values to train ANFIS classifier.