Update README.md
Dataset:

This database consists of two sets:   Main Leafs and Mask Leafs. 
Main Leafs: Main Leafs are 272 leaf images including 172 images of tomato leaves infected with both Alternaria and Leafminear and 100 healthy tomato leaves.
Mask Leafs: Mask Leafs are images 1 to 172 with Mask index. These are images that are the result of labeling the Main Leafs. In this dataset, Leafminear pest areas are labeled red, and Alternaria disease areas are labeled blue. 
The database is available in this page. kobra.nazari.cs@Gmail.com